refactor? Removed BlogDetail component from components/blog directory and moved it to pages/BlogDetail
