+++ src/router/AppRouter.jsx
@@ -8,7 +8,7 @@ import PrivateRouter from "./PrivateRouter";
 import Home from "../pages/Home";
 import Contact from "../pages/Contact";
 import Navbar from "../components/Navbar";
-import BlogDetail from "../components/Blog/BlogDetail";
+import BlogDetail from "../pages/BlogDetail";
 import Blog from "../pages/Blog";
 import Team from "../pages/Team";
 import Login from "../pages/Login";
